aggregate flow algorithms in separate module
