+++ src/algo/maximum_flow/ford_fulkerson.rs
@@ -1,15 +1,14 @@
 use std::{collections::VecDeque, ops::Sub};
 
 use crate::{
+    algo::{EdgeRef, PositiveMeasure},
     data::DataMap,
+    prelude::Direction,
     visit::{
         EdgeCount, EdgeIndexable, IntoEdges, IntoEdgesDirected, NodeCount, NodeIndexable, VisitMap,
         Visitable,
     },
 };
-
-use super::{EdgeRef, PositiveMeasure};
-use crate::prelude::Direction;
 
 fn residual_capacity<N>(
     network: N,
@@ -48,11 +47,12 @@ where
     }
 }
 
-fn bfs<N>(
+/// Tells whether there is an augmented path in the graph
+fn has_augmented_path<N>(
     network: N,
     source: N::NodeId,
     destination: N::NodeId,
-    vertex_levels: &mut [usize],
+    edge_to: &mut [Option<N::EdgeRef>],
     flows: &[N::EdgeWeight],
 ) -> bool
 where
@@ -63,10 +63,8 @@ where
     let mut queue = VecDeque::new();
     visited.visit(source);
     queue.push_back(source);
-    vertex_levels[NodeIndexable::to_index(&network, source)] = 1;
 
     while let Some(vertex) = queue.pop_front() {
-        let vertex_level = vertex_levels[NodeIndexable::to_index(&network, vertex)];
         let out_edges = network.edges_directed(vertex, Direction::Outgoing);
         let in_edges = network.edges_directed(vertex, Direction::Incoming);
         for edge in out_edges.chain(in_edges) {
@@ -75,51 +73,11 @@ where
             let residual_cap = residual_capacity(&network, edge, next, flows[edge_index]);
             if !visited.is_visited(&next) && (residual_cap > N::EdgeWeight::zero()) {
                 visited.visit(next);
-                vertex_levels[NodeIndexable::to_index(&network, next)] = vertex_level + 1;
-                queue.push_back(next);
-            }
-        }
-    }
-    let destination_level = vertex_levels[NodeIndexable::to_index(&network, destination)];
-    destination_level > 0
-}
-
-fn dfs<N>(
-    network: N,
-    source: N::NodeId,
-    destination: N::NodeId,
-    edge_to: &mut [Option<N::EdgeRef>],
-    vertex_levels: &[usize],
-    flows: &[N::EdgeWeight],
-) -> bool
-where
-    N: NodeCount + IntoEdgesDirected + NodeIndexable + EdgeIndexable + Visitable,
-    N::EdgeWeight: Sub<Output = N::EdgeWeight> + PositiveMeasure,
-{
-    let mut visited = network.visit_map();
-    let mut queue = VecDeque::new();
-    visited.visit(source);
-    queue.push_back(source);
-
-    while let Some(vertex) = queue.pop_front() {
-        let out_edges = network.edges_directed(vertex, Direction::Outgoing);
-        let in_edges = network.edges_directed(vertex, Direction::Incoming);
-        let curr_level = vertex_levels[NodeIndexable::to_index(&network, vertex)];
-        for edge in out_edges.chain(in_edges) {
-            let next = other_endpoint(&network, edge, vertex);
-            let edge_index: usize = EdgeIndexable::to_index(&network, edge.id());
-            let residual_cap = residual_capacity(&network, edge, next, flows[edge_index]);
-            let next_level = vertex_levels[NodeIndexable::to_index(&network, next)];
-            if next_level == curr_level + 1
-                && !visited.is_visited(&next)
-                && (residual_cap > N::EdgeWeight::zero())
-            {
-                visited.visit(next);
                 edge_to[NodeIndexable::to_index(&network, next)] = Some(edge);
                 if destination == next {
                     return true;
                 }
-                queue.push_front(next);
+                queue.push_back(next);
             }
         }
     }
@@ -149,13 +107,13 @@ where
     }
 }
 
-/// Dinic's (or Dinitz's) algorithm.
+/// \[Generic\] Ford-Fulkerson algorithm.
 ///
 /// Computes the [maximum flow][ff] of a weighted directed graph.
 ///
-/// Returns the maximum flow and also the computed edge flows.
+/// If it terminates, it returns the maximum flow and also the computed edge flows.
 ///
-/// [ff]: https://en.wikipedia.org/wiki/Dinic%27s_algorithm
+/// [ff]: https://en.wikipedia.org/wiki/Ford%E2%80%93Fulkerson_algorithm
 ///
 /// # Example
 /// ```rust
@@ -181,10 +139,10 @@ where
 ///    (4, 3, 7),
 ///    (4, 5, 4),
 /// ]);
-/// let (max_flow, _) = dinics(&graph, source, destination);
+/// let (max_flow, _) = ford_fulkerson(&graph, source, destination);
 /// assert_eq!(23, max_flow);
 /// ```
-pub fn dinics<N>(
+pub fn ford_fulkerson<N>(
     network: N,
     source: N::NodeId,
     destination: N::NodeId,
@@ -199,44 +157,39 @@ where
         + Visitable,
     N::EdgeWeight: Sub<Output = N::EdgeWeight> + PositiveMeasure,
 {
-    let mut levels = vec![0; network.node_count()];
     let mut edge_to = vec![None; network.node_count()];
     let mut flows = vec![N::EdgeWeight::zero(); network.edge_count()];
     let mut max_flow = N::EdgeWeight::zero();
-    while bfs(&network, source, destination, &mut levels, &flows) {
-        while dfs(&network, source, destination, &mut edge_to, &levels, &flows) {
-            let mut path_flow = N::EdgeWeight::max();
+    while has_augmented_path(&network, source, destination, &mut edge_to, &flows) {
+        let mut path_flow = N::EdgeWeight::max();
 
-            // Find the bottleneck capacity of the path
-            let mut vertex = destination;
-            let mut vertex_index = NodeIndexable::to_index(&network, vertex);
-            while let Some(edge) = edge_to[vertex_index] {
-                let edge_index = EdgeIndexable::to_index(&network, edge.id());
-                let residual_capacity =
-                    residual_capacity(&network, edge, vertex, flows[edge_index]);
-                // Minimum between the current path flow and the residual capacity.
-                path_flow = if path_flow > residual_capacity {
-                    residual_capacity
-                } else {
-                    path_flow
-                };
-                vertex = other_endpoint(&network, edge, vertex);
-                vertex_index = NodeIndexable::to_index(&network, vertex);
-            }
-
-            // Update the flow of each edge along the discovered path
-            let mut vertex = destination;
-            let mut vertex_index = NodeIndexable::to_index(&network, vertex);
-            while let Some(edge) = edge_to[vertex_index] {
-                let edge_index = EdgeIndexable::to_index(&network, edge.id());
-                flows[edge_index] =
-                    adjust_residual_flow(&network, edge, vertex, flows[edge_index], path_flow);
-                vertex = other_endpoint(&network, edge, vertex);
-                vertex_index = NodeIndexable::to_index(&network, vertex);
-            }
-            max_flow = max_flow + path_flow;
+        // Find the bottleneck capacity of the path
+        let mut vertex = destination;
+        let mut vertex_index = NodeIndexable::to_index(&network, vertex);
+        while let Some(edge) = edge_to[vertex_index] {
+            let edge_index = EdgeIndexable::to_index(&network, edge.id());
+            let residual_capacity = residual_capacity(&network, edge, vertex, flows[edge_index]);
+            // Minimum between the current path flow and the residual capacity.
+            path_flow = if path_flow > residual_capacity {
+                residual_capacity
+            } else {
+                path_flow
+            };
+            vertex = other_endpoint(&network, edge, vertex);
+            vertex_index = NodeIndexable::to_index(&network, vertex);
         }
-        levels = vec![0; network.node_count()];
+
+        // Update the flow of each edge along the path
+        let mut vertex = destination;
+        let mut vertex_index = NodeIndexable::to_index(&network, vertex);
+        while let Some(edge) = edge_to[vertex_index] {
+            let edge_index = EdgeIndexable::to_index(&network, edge.id());
+            flows[edge_index] =
+                adjust_residual_flow(&network, edge, vertex, flows[edge_index], path_flow);
+            vertex = other_endpoint(&network, edge, vertex);
+            vertex_index = NodeIndexable::to_index(&network, vertex);
+        }
+        max_flow = max_flow + path_flow;
     }
     (max_flow, flows)
 }
