+++ src/algo/maximum_flow/mod.rs
@@ -1,0 +1,5 @@
+mod dinics;
+mod ford_fulkerson;
+
+pub use dinics::dinics;
+pub use ford_fulkerson::ford_fulkerson;
